requirements.txt
# Core
python-dotenv==1.0.1
requests==2.31.0

# LLM & Embeddings (OpenAI) - Use version compatible with httpx
openai==1.14.0

# UI
streamlit==1.37.0

# Data Processing
pandas==2.1.4
tiktoken==0.7.0

# HTTP client (pin to avoid proxy issues)
httpx==0.27.0

# Testing & Development
pytest==8.0.0
pytest-cov==4.1.0
black==24.1.0
flake8==7.0.0
